+++ src/Dom/DocumentBuilder.ts
@@ -1,14 +1,8 @@
-import {AbstractGuiStubBuilder} from "./index";
+import {AbstractDomBuilder} from "./index";
 
-export class DocumentBuilder extends AbstractGuiStubBuilder {
+export class DocumentBuilder extends AbstractDomBuilder {
 
     private fs = require("fs");
-
-    public setContentFromUrl(url: string): this {
-        let res = this.request("GET", url);
-        this.setSource(res.getBody());
-        return this;
-    }
 
     public setSourceFromFile(filename: string): this {
         let content: string = this.fs.readFileSync(filename);