+++ src/Dom/LocationBuilder.ts
@@ -1,0 +1,8 @@
+import {AbstractDomBuilder} from "./AbstractDomBuilder";
+
+export class LocationBuilder extends AbstractDomBuilder {
+
+    public getMock(): Location {
+        return this.createNewWindow().location;
+    }
+}