+++ src/Mock/index.ts
@@ -1,5 +1,2 @@
-export {AbstractGuiStubBuilder} from "./AbstractGuiStubBuilder";
 export {BuilderInterface} from "./BuilderInterface";
-export {DocumentBuilder} from "./DocumentBuilder";
 export {ObjectBuilder} from "./ObjectBuilder";
-export {WindowBuilder} from "./WindowBuilder";