+++ tests/Dom/WindowBuilderTest.ts
@@ -1,10 +1,10 @@
-import {AbstractUnitTestCase} from "../../dist/AbstractUnitTestCase";
-import {WindowBuilder} from "../../dist/Mock/index";
+import {WindowBuilder} from "../../dist/Dom/index";
+import {AbstractDomTestCase} from "../../dist/AbstractDomTestCase";
 
-export class WindowBuilderTest extends AbstractUnitTestCase {
+export class WindowBuilderTest extends AbstractDomTestCase {
 
     public testStubWindowFromUrl() {
-        let builder: WindowBuilder = new WindowBuilder();
+        let builder: WindowBuilder = this.getWindowBuilder();
         let window: Window = builder.setLocation("http://example.com", true).getMock();
         this.assertInstanceOf(Window, window);
         this.assertEquals("example.com", window.location.host);